+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### 🎉 Added
 
-- Live update the conception graph on each file save
+- Live update the design graph on each file save
 
 ### 🦋 Changed
 
@@ -26,7 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### 🎉 Added
 
 - Provide include/ignore file patterns configurability
-- Provide conception graph layout configurability
+- Provide design graph layout configurability
 
 ### 🦋 Changed