+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### 🚑 Fixed
 
 - Fix graph rendering (copy rendering templates during compilation)
+- Rename "Conception" into "Design" for more accuracy
 
 ## [0.2.0] - 2020-04-22